+++ app/helpers/agilibox/pagination_helper.rb
@@ -3,4 +3,12 @@ module Agilibox::PaginationHelper
     options = {theme: "twitter-bootstrap-3"}.merge(options)
     super(objects, options).gsub(/>(\s+)</, "><").html_safe
   end
+
+  def pagination_infos(collection)
+    content_tag(:p, class: "pagination-infos") { page_entries_info(collection) }
+  end
+
+  def pagination_and_infos(collection)
+    paginate(collection) + pagination_infos(collection)
+  end
 end